Fixed the framebuffer code, and some others

+++ LearningEngine/src/Graphics/Renderer/API/Renderer.h
@@ -21,13 +21,13 @@ public:
 	virtual RenderContext* GetContext() = 0;
 	virtual std::string GetVersionString() = 0;
 	virtual void Clear(const glm::vec4& color) = 0;
-	virtual void Resize(const float width, const float height) = 0;
+	virtual void Resize(const int width, const int height) = 0;
 	virtual void Present() = 0;
 	virtual void UseDepthTesting(bool enable) = 0;
 	virtual void BindViewport() = 0;
-	virtual void SetViewportSize(float width, float height) = 0;
+	virtual void SetViewportSize(int width, int height) = 0;
 	virtual glm::vec2 GetViewportSize() = 0;
-	virtual void SetViewportPosition(float x, float y) = 0;
+	virtual void SetViewportPosition(int x, int y) = 0;
 	virtual glm::vec2 GetViewportPosition() = 0;
 
 	static API GetAPI() { return s_API; }
