+++ LearningEngine/src/Platform/OpenGL/OGLRenderer.h
@@ -18,17 +18,17 @@ public:
 	virtual RenderContext* GetContext() override;
 	virtual std::string GetVersionString() override;
 	virtual void Clear(const glm::vec4& color) override;
-	virtual void Resize(const float width, const float height) override;
+	virtual void Resize(const int width, const int height) override;
 	virtual void Present() override;
 	virtual void UseDepthTesting(bool enable) override;
 	virtual void BindViewport() override;
-	virtual void SetViewportSize(float width, float height) override;
+	virtual void SetViewportSize(int width, int height) override;
 	virtual glm::vec2 GetViewportSize() override;
-	virtual void SetViewportPosition(float x, float y) override;
+	virtual void SetViewportPosition(int x, int y) override;
 	virtual glm::vec2 GetViewportPosition() override;
 
 private:
-	glm::vec4 m_Viewport;
+	glm::ivec4 m_Viewport;
 	WindowsWindow m_Window;
 	RenderContext* m_Context;
 };
